refact: TBTI질문 관리기능을 TbtiManagerService로 분리
